+++ test/common.py
@@ -276,7 +276,7 @@ def run_train_translate(train_params: str,
         if 'train_source_factors' in data:
             prepare_params += _TRAIN_WITH_FACTORS_COMMON.format(source_factors=" ".join(data['train_source_factors']))
 
-        if '--weight-tying' in train_params and '--weight-tying-type src_trg' in train_params:
+        if '--weight-tying-type src_trg' in train_params or '--weight-tying-type src_trg_softmax' in train_params:
             prepare_params += ' --shared-vocab'
 
         logger.info("Preparing data with parameters %s.", prepare_params)